requirements.txt
python-dotenv
pyrofork
TgCrypto-pyrofork
aiohttp